(PC-18245) feat(shareApp): Refactor to respect architecture standard
