+++ src/features/auth/signup/AccountCreated/AccountCreated.tsx
@@ -8,9 +8,9 @@ import {
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { useUserProfileInfo } from 'features/profile/api'
 import { useShareAppContext } from 'features/shareApp/context/ShareAppWrapper'
+import { ShareAppModal } from 'features/shareApp/helpers/shareAppModalInformations'
 import { campaignTracker, CampaignEvents } from 'libs/campaign'
 import { BatchEvent, BatchUser } from 'libs/react-native-batch'
-import { ShareAppModal } from 'libs/share/shareApp/shareAppModalInformations'
 import IlluminatedSmileyAnimation from 'ui/animations/lottie_illuminated_smiley.json'
 import { ButtonPrimaryWhite } from 'ui/components/buttons/ButtonPrimaryWhite'
 import { TouchableLink } from 'ui/components/touchableLink/TouchableLink'
