+++ src/features/auth/signup/AccountCreated/tests/AccountCreated.native.test.tsx
@@ -9,8 +9,8 @@ import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { navigateFromRef } from 'features/navigation/navigationRef'
 import { useUserProfileInfo } from 'features/profile/api'
 import { ShareAppWrapper } from 'features/shareApp/context/ShareAppWrapper'
+import { ShareAppModal } from 'features/shareApp/helpers/shareAppModalInformations'
 import { BatchUser } from 'libs/react-native-batch'
-import { ShareAppModal } from 'libs/share/shareApp/shareAppModalInformations'
 import { render, fireEvent } from 'tests/utils'
 
 import { AccountCreated } from '../AccountCreated'
