+++ src/features/cheatcodes/pages/NavigationShareApp/NavigationShareApp.tsx
@@ -3,8 +3,8 @@ import { ScrollView } from 'react-native'
 import styled from 'styled-components/native'
 
 import { Row } from 'features/cheatcodes/components/Row'
-import { ShareAppModal } from 'libs/share/shareApp/shareAppModalInformations'
-import { ShareAppModalNew } from 'libs/share/shareApp/ShareAppModalNew'
+import { ShareAppModalNew } from 'features/shareApp/components/ShareAppModalNew'
+import { ShareAppModal } from 'features/shareApp/helpers/shareAppModalInformations'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
 import { PageHeader } from 'ui/components/headers/PageHeader'
 import { useModal } from 'ui/components/modals/useModal'
