+++ src/features/identityCheck/pages/confirmation/BeneficiaryAccountCreated.tsx
@@ -17,9 +17,9 @@ import { useUserProfileInfo } from 'features/profile/api'
 import { isUserUnderageBeneficiary } from 'features/profile/utils'
 import { useMaxPrice } from 'features/search/utils/useMaxPrice'
 import { useShareAppContext } from 'features/shareApp/context/ShareAppWrapper'
+import { ShareAppModal } from 'features/shareApp/helpers/shareAppModalInformations'
 import { formatPriceInEuroToDisplayPrice } from 'libs/parsers'
 import { BatchEvent, BatchUser } from 'libs/react-native-batch'
-import { ShareAppModal } from 'libs/share/shareApp/shareAppModalInformations'
 import TutorialPassLogo from 'ui/animations/tutorial_pass_logo.json'
 import { AnimatedProgressBar } from 'ui/components/bars/AnimatedProgressBar'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
