+++ src/features/shareApp/context/ShareAppWrapper.tsx
@@ -1,7 +1,7 @@
 import React, { memo, useCallback, useContext, useMemo, useState } from 'react'
 
-import { ShareAppModal } from 'libs/share/shareApp/shareAppModalInformations'
-import { ShareAppModalNew } from 'libs/share/shareApp/ShareAppModalNew'
+import { ShareAppModalNew } from 'features/shareApp/components/ShareAppModalNew'
+import { ShareAppModal } from 'features/shareApp/helpers/shareAppModalInformations'
 import { useModal } from 'ui/components/modals/useModal'
 
 interface ShareAppContextValue {
