+++ src/features/shareApp/helpers/shareAppModalInformations.ts
